Adicionado paginação

+++ client/src/pages/Books/index.js
@@ -9,22 +9,32 @@ export default function Books(){
 
 
     const [books, setBooks] = useState([]);
+    const [page, setPage] = useState(1);
+
 
     const userName = localStorage.getItem('userName');
 
     const accessToken = localStorage.getItem('accessToken');
+
+    const authorization = {
+        headers:{
+            Authorization: `Bearer ${accessToken}`
+        }
+    }
+
     const history = useHistory();
 
 
     useEffect(() => {
-        api.get('api/Book/v1/asc/20/1', {
-            headers:{
-                Authorization: `Bearer ${accessToken}`
-            }
-        }).then(response => {
-            setBooks(response.data.list)
-        })
+        fetchMoreBooks();
     }, [accessToken]);
+
+    async function fetchMoreBooks(){
+        const response = await api.get(`api/Book/v1/asc/4/${page}`, authorization);
+            setBooks([...books, ...response.data.list])
+            setPage(page + 1);
+        
+    }
 
     async function logout(){
         try {
@@ -52,11 +62,7 @@ export default function Books(){
 
     async function deleteBook(id){
         try {
-            await api.delete(`api/Book/v1/${id}`, {
-                headers:{
-                    Authorization: `Bearer ${accessToken}`
-                }
-            });
+            await api.delete(`api/Book/v1/${id}`, authorization);
             setBooks(books.filter(book => book.id !== id))
         } catch (error) {
             alert("Houve um erro ao deletar o livro, tente novamente mais tarde!");
@@ -95,6 +101,7 @@ export default function Books(){
                 </li>
                ))}
             </ul>
+            <button className="button" onClick={fetchMoreBooks} type="button">Loading More</button>
         </div>
     );
 }
